Remove unused build artifacts and cleanup generated files
