+++ src/game/constants.js
@@ -1,0 +1,10 @@
+export const CONTROL_DIAMETER = 110;
+export const LOOK_RADIUS = 90;
+export const DEADZONE = 0.10;
+
+export const MAP_W = 16;
+export const MAP_H = 16;
+export const FOV = Math.PI / 3;
+export const MAX_DEPTH = 20;
+export const STEP = 0.02;
+export const PITCH_LIMIT = Math.PI * 80 / 180;
