+++ src/game/map.js
@@ -1,0 +1,35 @@
+import { MAP_W, MAP_H } from './constants.js';
+
+export const BASE_MAP = [
+  1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,
+  1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,
+  1,0,0,0,0,1,1,0,0,0,0,0,0,0,0,1,
+  1,0,0,0,0,1,0,0,0,1,1,1,0,1,0,1,
+  1,0,0,1,0,1,0,0,0,0,0,1,0,1,0,1,
+  1,0,0,1,0,1,0,0,0,0,0,1,0,1,0,1,
+  1,0,0,1,0,1,0,0,0,1,0,1,0,0,0,1,
+  1,0,0,1,0,1,0,0,0,1,0,0,0,0,0,1,
+  1,0,0,0,0,0,0,0,0,1,0,0,0,0,0,1,
+  1,0,0,0,0,0,0,0,0,1,0,0,0,0,0,1,
+  1,0,0,1,0,1,0,0,0,1,0,1,0,1,0,1,
+  1,0,0,1,0,1,0,0,0,1,0,1,0,1,0,1,
+  1,0,0,1,0,0,0,0,0,1,0,1,0,1,0,1,
+  1,0,0,1,0,0,0,0,0,0,0,1,0,0,0,1,
+  1,0,0,0,0,0,0,0,0,0,0,0,0,0,2,1,
+  1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,
+];
+
+export function idx(x, y){
+  return y * MAP_W + x;
+}
+
+export function tileAt(px, py){
+  const gx = Math.floor(px);
+  const gy = Math.floor(py);
+  if(gx < 0 || gy < 0 || gx >= MAP_W || gy >= MAP_H) return 1;
+  return BASE_MAP[idx(gx, gy)];
+}
+
+export function isWall(px, py){
+  return tileAt(px, py) === 1;
+}
